now the app uses context
