+++ src/components/TodoList.tsx
@@ -1,51 +1,34 @@
-import { FC, memo } from 'react';
-import { Todo } from '../types/Todo';
+import { FC, memo, useContext } from 'react';
+import { GlobalContext } from '../contexts/GlobalContext';
 import { TodoComponent } from './Todo';
 
-interface Props {
-  visibleTodos: Todo[],
-  tempTodo: Todo | null,
-  isLoading: number[],
-  onDelete: (id: number) => void,
-  onToggle: (id: number) => void,
-  onRename: (newTitle: string, id: number) => void
-}
+export const TodoList: FC = memo(
+  () => {
+    const {
+      visibleTodos,
+      temporaryTodo,
+    } = useContext(GlobalContext);
 
-export const TodoList: FC<Props> = memo(
-  ({
-    visibleTodos,
-    tempTodo,
-    isLoading,
-    onDelete,
-    onToggle,
-    onRename,
-  }) => (
-    <section className="todoapp__main" data-cy="TodoList">
-      {
-        visibleTodos
-          .map((todo) => (
+    return (
+      <section className="todoapp__main" data-cy="TodoList">
+        {
+          visibleTodos
+            .map((todo) => (
+              <TodoComponent
+                key={todo.id}
+                todo={todo}
+              />
+            ))
+        }
+
+        {
+          temporaryTodo && (
             <TodoComponent
-              key={todo.id}
-              todo={todo}
-              isLoading={isLoading}
-              onDelete={onDelete}
-              onToggle={onToggle}
-              onRename={onRename}
+              todo={temporaryTodo}
             />
-          ))
-      }
-
-      {
-        tempTodo && (
-          <TodoComponent
-            todo={tempTodo}
-            isLoading={isLoading}
-            onDelete={onDelete}
-            onToggle={onToggle}
-            onRename={onRename}
-          />
-        )
-      }
-    </section>
-  ),
+          )
+        }
+      </section>
+    );
+  },
 );
