+++ src/components/footer.tsx
@@ -1,40 +1,54 @@
-import { FC, memo } from 'react';
+import { FC, memo, useContext } from 'react';
+import { DeleteContext } from '../contexts/DeleteContext';
+import { GlobalContext } from '../contexts/GlobalContext';
 import { Filter } from './Filter';
 
-interface Props {
-  numberOfItems: number,
-  filterStatus: string,
-  setFilterStatus: React.Dispatch<React.SetStateAction<string>>,
-  onClearCompleted: () => void,
-  isClearCompletedHidden: boolean,
-}
+export const Footer: FC = memo(
+  () => {
+    const {
+      visibleTodos,
+      filterStatus,
+      setFilterStatus,
+    } = useContext(GlobalContext);
 
-export const Footer: FC<Props> = memo(
-  ({
-    numberOfItems,
-    filterStatus,
-    isClearCompletedHidden,
-    setFilterStatus,
-    onClearCompleted,
-  }) => (
-    <footer className="todoapp__footer" data-cy="Footer">
-      <span className="todo-count" data-cy="todosCounter">
-        {`${numberOfItems} items left`}
-      </span>
+    const {
+      onClearCompleted,
+      isClearCompletedHidden,
+    } = useContext(DeleteContext);
 
-      <Filter filterStatus={filterStatus} onFilter={setFilterStatus} />
+    if (visibleTodos.length === 0) {
+      return <></>;
+    }
 
-      <button
-        data-cy="ClearCompletedButton"
-        type="button"
-        className="todoapp__clear-completed"
-        onClick={onClearCompleted}
-        style={{
-          visibility: !isClearCompletedHidden ? 'hidden' : 'visible',
-        }}
+    const unFinishedTasks = visibleTodos.reduce((acc, task) => {
+      return task.completed
+        ? acc
+        : acc + 1;
+    }, 0);
+
+    return (
+      <footer
+        className="todoapp__footer"
+        data-cy="Footer"
       >
-        Clear completed
-      </button>
-    </footer>
-  ),
+        <span className="todo-count" data-cy="todosCounter">
+          {`${unFinishedTasks} items left`}
+        </span>
+
+        <Filter filterStatus={filterStatus} onFilter={setFilterStatus} />
+
+        <button
+          data-cy="ClearCompletedButton"
+          type="button"
+          className="todoapp__clear-completed"
+          onClick={onClearCompleted}
+          style={{
+            visibility: !isClearCompletedHidden ? 'hidden' : 'visible',
+          }}
+        >
+          Clear completed
+        </button>
+      </footer>
+    );
+  },
 );
